requirements.txt
Flask==2.0.1
google-cloud-bigquery==2.30.0
gunicorn==20.1.0
Werkzeug==2.0.1
requests==2.26.0